Fixing Code Review Findings and Reorganizing File structure
