+++ bundles/org.openhab.binding.sonnen/src/main/java/org/openhab/binding/sonnen/internal/utilities/Helper.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.sonnen.utilities;
+package org.openhab.binding.sonnen.internal.utilities;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
