+++ bundles/org.openhab.binding.sonnen/src/main/resources/OH-INF/i18n/sonnen.properties
@@ -13,7 +13,7 @@ thing-type.sonnen.sonnenbattery.description = Thing for sonnen battery connectio
 thing-type.config.sonnen.sonnenbattery.hostIP.label = IP Address
 thing-type.config.sonnen.sonnenbattery.hostIP.description = Please add the IP Address of your sonnen battery
 thing-type.config.sonnen.sonnenbattery.refreshInterval.label = Refresh Interval
-thing-type.config.sonnen.sonnenbattery.refreshInterval.description = How often in seconds the sonnen battery should schedule a refresh after a channel is linked to an item. Valid input is 0 - 1000
+thing-type.config.sonnen.sonnenbattery.refreshInterval.description = How often  the sonnen battery in seconds should schedule a refresh after a channel is linked to an item. Valid input is 0 - 1000
 
 # channel types
